saju-engine/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ephem==4.1.6
slowapi==0.1.9
python-dotenv==1.0.1
pydantic==2.10.3
